+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Configuration.php
@@ -551,7 +551,7 @@ class Configuration implements ConfigurationInterface
                         ->scalarNode('version_strategy')->defaultNull()->end()
                         ->scalarNode('version')->defaultNull()->end()
                         ->scalarNode('version_format')->defaultValue('%%s?%%s')->end()
-                        ->scalarNode('manifest_path')->defaultNull()->end()
+                        ->scalarNode('json_manifest_path')->defaultNull()->end()
                         ->scalarNode('base_path')->defaultValue('')->end()
                         ->arrayNode('base_urls')
                             ->requiresAtLeastOneElement()
@@ -570,9 +570,15 @@ class Configuration implements ConfigurationInterface
                     ->end()
                     ->validate()
                         ->ifTrue(function ($v) {
-                            return isset($v['version_strategy']) && $v['version_strategy'] === 'json_manifest' && !isset($v['manifest_path']);
+                            return isset($v['version_strategy']) && isset($v['json_manifest_path']);
                         })
-                        ->thenInvalid('You must configure the "manifest_path" key in order to use the "json_manifest" "version_strategy".')
+                        ->thenInvalid('You cannot use both "version_strategy" and "json_manifest_path" at the same time under "assets".')
+                    ->end()
+                    ->validate()
+                        ->ifTrue(function ($v) {
+                            return isset($v['version']) && isset($v['json_manifest_path']);
+                        })
+                        ->thenInvalid('You cannot use both "version" and "json_manifest_path" at the same time under "assets".')
                     ->end()
                     ->fixXmlConfig('package')
                     ->children()
@@ -589,7 +595,7 @@ class Configuration implements ConfigurationInterface
                                         ->end()
                                     ->end()
                                     ->scalarNode('version_format')->defaultNull()->end()
-                                    ->scalarNode('manifest_path')->defaultNull()->end()
+                                    ->scalarNode('json_manifest_path')->defaultNull()->end()
                                     ->scalarNode('base_path')->defaultValue('')->end()
                                     ->arrayNode('base_urls')
                                         ->requiresAtLeastOneElement()
@@ -608,9 +614,15 @@ class Configuration implements ConfigurationInterface
                                 ->end()
                                 ->validate()
                                     ->ifTrue(function ($v) {
-                                        return isset($v['version_strategy']) && $v['version_strategy'] === 'json_manifest' && !isset($v['manifest_path']);
+                                        return isset($v['version_strategy']) && isset($v['json_manifest_path']);
                                     })
-                                    ->thenInvalid('You must configure the "manifest_path" key in order to use the "json_manifest" "version_strategy".')
+                                    ->thenInvalid('You cannot use both "version_strategy" and "json_manifest_path" at the same time under "assets" packages.')
+                                ->end()
+                                ->validate()
+                                    ->ifTrue(function ($v) {
+                                        return isset($v['version']) && isset($v['json_manifest_path']);
+                                    })
+                                    ->thenInvalid('You cannot use both "version" and "json_manifest_path" at the same time under "assets" packages.')
                                 ->end()
                             ->end()
                         ->end()